cleaned code, 3 songs added, handled 0-sec-left situations
This is probably the last time I will continue coding this game. I might call it a day now. It took me so much time to try and figure out how to prevent memory leaks from SDL_Surface and SDL_Texture. Unfortunately, after so many attempts and repeatedly replacing large amounts of code, I haven't been successful. This game will be left with that issue. If there's anyone reading this who has lots of time, could you do me a favor and fix the memory leaking issue? To describe the issue, if I don't use SDL_FreeSurface, I only rarely see flickers of some texture on the screen. But if I use SDL_FreeSurface, it happens very frequently. Also, there is a very small chance that you might get the car to flip repetitively, which is of course undesirable. I will not come back to fix those issues, at least not in the meantime. Besides, I should mention that my game is not very optimized, as there is still a lot of room for improvement.